+++ server/src-lib/Hasura/SQL/Types.hs
@@ -341,6 +341,7 @@ data PGScalarType
   | PGFloat
   | PGDouble
   | PGNumeric
+  | PGMoney
   | PGBoolean
   | PGChar
   | PGVarchar
@@ -371,6 +372,7 @@ instance ToSQL PGScalarType where
     PGFloat       -> "real"
     PGDouble      -> "float8"
     PGNumeric     -> "numeric"
+    PGMoney       -> "money"
     PGBoolean     -> "boolean"
     PGChar        -> "character"
     PGVarchar     -> "varchar"
@@ -419,6 +421,8 @@ textToPGScalarType t = case t of
   "numeric"                  -> PGNumeric
   "decimal"                  -> PGNumeric
 
+  "money"                    -> PGMoney
+
   "boolean"                  -> PGBoolean
   "bool"                     -> PGBoolean
 
@@ -462,6 +466,7 @@ pgTypeOid PGBigSerial   = PTI.int8
 pgTypeOid PGFloat       = PTI.float4
 pgTypeOid PGDouble      = PTI.float8
 pgTypeOid PGNumeric     = PTI.numeric
+pgTypeOid PGMoney       = PTI.numeric
 pgTypeOid PGBoolean     = PTI.bool
 pgTypeOid PGChar        = PTI.char
 pgTypeOid PGVarchar     = PTI.varchar
@@ -488,6 +493,7 @@ isNumType :: PGScalarType -> Bool
 isNumType PGFloat   = True
 isNumType PGDouble  = True
 isNumType PGNumeric = True
+isNumType PGMoney   = True
 isNumType ty        = isIntegerType ty
 
 stringTypes :: [PGScalarType]
@@ -517,6 +523,7 @@ isBigNum = \case
   PGBigSerial -> True
   PGNumeric   -> True
   PGDouble    -> True
+  PGMoney     -> True
   _           -> False
 
 geoTypes :: [PGScalarType]