+++ CHANGELOG.md
@@ -388,6 +388,7 @@ The order, collapsed state of columns and rows limit is now persisted across pag
 - console: surround string type column default value with quotes (close #4371) (#4423)
 - console: add undefined check to fix error (close #4444) (#4445)
 - docs: add One-Click Render deployment guide (close #3683) (#4209)
+- server: add support for `_inc` on `real`, `double`, `numeric` and `money` (fix #3573)
 - server: reserved keywords in column references break parser (fix #3597) #3927
 - server: fix postgres specific error message that exposed database type on invalid query parameters (#4294)
 - server: manage inflight events when HGE instance is gracefully shutdown (close #3548)